backend/requirements/base.txt
# TeamTrack – Base requirements
# Python 3.10+

Django>=4.2,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
